refactor: Improved React file structure refactor: Moved AnchorButton to HTMLComponents
